requirements.txt
fastapi
uvicorn
pandas
numpy
statsmodels
python-dotenv